% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_config.R
\name{as.list.MgrConfig}
\alias{as.list.MgrConfig}
\title{Convert a MgrConfig to a plain named list}
\usage{
\method{as.list}{MgrConfig}(x, ...)
}
\arguments{
\item{x}{A \code{MgrConfig} object.}

\item{...}{Ignored.}
}
\description{
Convert a MgrConfig to a plain named list
}
