% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rope.R
\name{broadcat}
\alias{broadcat}
\title{Broadcast and concatenate a list of tensors}
\usage{
broadcat(tensors, dim = -1L)
}
\arguments{
\item{tensors}{List of tensors to broadcast and concatenate}

\item{dim}{Dimension along which to concatenate (default: \code{-1L})}
}
\value{
Concatenated tensor after broadcasting
}
\description{
Equivalent to the \code{broadcat} utility from tortoise-tts.
}
\keyword{internal}
