% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rope.R
\name{default_val}
\alias{default_val}
\title{Return value or default fallback}
\usage{
default_val(val, d)
}
\arguments{
\item{val}{Value to check}

\item{d}{Default value to return if \code{val} is \code{NULL}}
}
\value{
\code{val} if not \code{NULL}, otherwise \code{d}
}
\description{
Return value or default fallback
}
\keyword{internal}
