% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{disk_tensor}
\alias{disk_tensor}
\title{A tensor backed by a memory-mapped file on disk}
\value{
A \code{DiskTensor} object.
}
\description{
Uses R's \code{bigmemory} or base file I/O for storage and provides a torch tensor
interface for seamless integration. This enables storing tensors larger than
available RAM by using disk space.
}
\note{
R6 doesn't support \code{[} overloading directly; use \verb{$get()} instead.
}
\section{Methods}{

\describe{
\item{\code{flush()}}{Flush changes to disk}
}
}

\examples{
\dontrun{
# Create a disk-backed tensor
dt <- disk_tensor$new(c(1000L, 128L), torch_float(), "/tmp/my_tensor.mmap")

# Write data (automatically persists to disk)
dt[1:100, ..] <- torch_randn(100L, 128L)

# Read data
x <- dt[1:50, ..]$tensor

# Flush changes to disk
dt$flush()
}

}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-DiskTensor-initialize}{\code{DiskTensor$new()}}
    \item \href{#method-DiskTensor-tensor}{\code{DiskTensor$tensor()}}
    \item \href{#method-DiskTensor-get}{\code{DiskTensor$get()}}
    \item \href{#method-DiskTensor-set}{\code{DiskTensor$set()}}
    \item \href{#method-DiskTensor-flush}{\code{DiskTensor$flush()}}
    \item \href{#method-DiskTensor-nbytes}{\code{DiskTensor$nbytes()}}
    \item \href{#method-DiskTensor-clone}{\code{DiskTensor$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiskTensor-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-DiskTensor-initialize}{}}}
\subsection{\code{DiskTensor$new()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiskTensor$new(shape, dtype, path, cleanup = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{shape}}{Integer vector. Shape of the tensor.}
      \item{\code{dtype}}{torch.dtype. Data type.}
      \item{\code{path}}{Character. File path.}
      \item{\code{cleanup}}{Logical. Auto-cleanup on GC.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiskTensor-tensor"></a>}}
\if{latex}{\out{\hypertarget{method-DiskTensor-tensor}{}}}
\subsection{\code{DiskTensor$tensor()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiskTensor$tensor()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A torch tensor.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiskTensor-get"></a>}}
\if{latex}{\out{\hypertarget{method-DiskTensor-get}{}}}
\subsection{\code{DiskTensor$get()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiskTensor$get(indices)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{indices}}{List of indices or slices.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A torch tensor.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiskTensor-set"></a>}}
\if{latex}{\out{\hypertarget{method-DiskTensor-set}{}}}
\subsection{\code{DiskTensor$set()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiskTensor$set(indices, value)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{indices}}{List of indices.}
      \item{\code{value}}{Tensor. Value to write.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiskTensor-flush"></a>}}
\if{latex}{\out{\hypertarget{method-DiskTensor-flush}{}}}
\subsection{\code{DiskTensor$flush()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiskTensor$flush()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Invisible flush to disk
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiskTensor-nbytes"></a>}}
\if{latex}{\out{\hypertarget{method-DiskTensor-nbytes}{}}}
\subsection{\code{DiskTensor$nbytes()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiskTensor$nbytes()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Integer. Size in bytes.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiskTensor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiskTensor-clone}{}}}
\subsection{\code{DiskTensor$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiskTensor$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
