% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{.canonical_mask}
\alias{.canonical_mask}
\title{Canonical mask helper for attention}
\usage{
.canonical_mask(
  mask,
  mask_name,
  other_type,
  other_name,
  target_type,
  check_other = TRUE
)
}
\description{
Canonical mask helper for attention
}
\keyword{internal}
