% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardhat_regressor.R
\name{.concat_first_axis}
\alias{.concat_first_axis}
\title{Concatenate a list of arrays along their first axis.}
\usage{
.concat_first_axis(arrays)
}
\arguments{
\item{arrays}{List of matrices or arrays with compatible dimensions.}
}
\value{
A single concatenated array.
}
\description{
Concatenate a list of arrays along their first axis.
}
\keyword{internal}
