% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rope.R
\name{.default}
\alias{.default}
\title{Return value or default fallback}
\usage{
.default(val, d)
}
\arguments{
\item{val}{Value to check}

\item{d}{Default value}
}
\value{
\code{val} if not \code{NULL}, otherwise \code{d}.
}
\description{
Return value or default fallback
}
\keyword{internal}
