% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kv-cache.R
\name{.element_size}
\alias{.element_size}
\title{Compute the number of bytes per element for common torch dtypes.}
\usage{
.element_size(tensor)
}
\arguments{
\item{tensor}{A torch tensor.}
}
\value{
\code{integer(1)} Bytes per element.
}
\description{
Compute the number of bytes per element for common torch dtypes.
}
\keyword{internal}
