% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attention.R
\docType{data}
\name{.flash_attn3_enabled}
\alias{.flash_attn3_enabled}
\title{Global flag for Flash Attention 3 usage}
\format{
An object of class \code{logical} of length 1.
}
\usage{
.flash_attn3_enabled
}
\description{
Global flag for Flash Attention 3 usage
}
\keyword{internal}
