% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{.load_tensor_from_disk}
\alias{.load_tensor_from_disk}
\title{Load tensor from disk (simplified implementation)}
\usage{
.load_tensor_from_disk(path, shape, dtype)
}
\description{
TODO swith to safetensors::safe_load_file()
}
\keyword{internal}
