% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{.make_buffer_key}
\alias{.make_buffer_key}
\title{Create a unique key for buffer pooling}
\usage{
.make_buffer_key(shape, dtype)
}
\arguments{
\item{shape}{Integer vector. Tensor shape.}

\item{dtype}{torch.dtype. Tensor dtype.}
}
\value{
Character. Unique key string.
}
\description{
Create a unique key for buffer pooling
}
\keyword{internal}
