% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_config.R
\name{.mgrcfg_update}
\alias{.mgrcfg_update}
\title{Update a MgrConfig list in place (returns updated copy)}
\usage{
.mgrcfg_update(x, updates)
}
\arguments{
\item{x}{A \code{MgrConfig} or named list.}

\item{updates}{A named list or \code{MgrConfig} with new values.}
}
\value{
Updated \code{MgrConfig}.
}
\description{
Update a MgrConfig list in place (returns updated copy)
}
\keyword{internal}
