% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{.none_or_dtype}
\alias{.none_or_dtype}
\title{Check if value is NULL or extract dtype from tensor}
\usage{
.none_or_dtype(x)
}
\description{
Check if value is NULL or extract dtype from tensor
}
\keyword{internal}
