% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardhat_regressor.R
\name{.split_first_axis}
\alias{.split_first_axis}
\title{Split an array along its first axis into roughly equal chunks.}
\usage{
.split_first_axis(arr, n)
}
\arguments{
\item{arr}{A matrix or array.}

\item{n}{\code{integer(1)} Number of chunks.}
}
\value{
A list of sub-arrays.
}
\description{
Split an array along its first axis into roughly equal chunks.
}
\keyword{internal}
