% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rope.R
\name{exists_val}
\alias{exists_val}
\title{Check if a value exists (is not NULL)}
\usage{
exists_val(val)
}
\arguments{
\item{val}{Value to check}
}
\value{
\code{TRUE} if \code{val} is not \code{NULL}, \code{FALSE} otherwise
}
\description{
Check if a value exists (is not NULL)
}
\keyword{internal}
