% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{format.OffloadReason}
\alias{format.OffloadReason}
\title{Format OffloadReason for display}
\usage{
\method{format}{OffloadReason}(x, ...)
}
\arguments{
\item{x}{An OffloadReason object}

\item{...}{Additional arguments (unused)}
}
\description{
Format OffloadReason for display
}
\keyword{internal}
