% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-coalescing operator for R}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Value to check}

\item{y}{Default value}
}
\value{
x if not NULL, else y
}
\description{
Null-coalescing operator for R
}
\keyword{internal}
