% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_config.R
\name{inference_config}
\alias{inference_config}
\title{Create an InferenceConfig with default settings}
\usage{
inference_config(...)
}
\arguments{
\item{...}{Arguments forwarded to \code{InferenceConfig$new()}.}
}
\value{
An \code{InferenceConfig} object.
}
\description{
Convenience alias for \code{InferenceConfig$new()}.
}
