% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{memory_estimator}
\alias{memory_estimator}
\title{MemoryEstimator: Estimate peak activation memory for inference components}
\usage{
memory_estimator
}
\description{
Estimates peak inference memory requirements for different attention-based components.
Peak inference memory refers to the maximum amount of memory (typically GPU memory) used
during the inference phase of a model.
}
\details{
The coefficients and intercepts for each component are derived through memory profiling
and regression using the model:
\deqn{
c_1 \cdot batch\_size + c_2 \cdot seq\_len + c_3 \cdot (batch\_size \times seq\_len) + intercept
}

Memory profiling was conducted using float32 without automatic mixed precision (AMP).
When using AMP, actual memory usage will be lower than the estimates.
}
\keyword{internal}
