% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{nn_ic_learning}
\alias{nn_ic_learning}
\title{Dataset-wise in-context learning.}
\usage{
nn_ic_learning(
  max_classes,
  out_dim,
  d_model,
  num_blocks,
  nhead,
  dim_feedforward,
  dropout = 0,
  activation = "gelu",
  norm_first = TRUE,
  bias_free_ln = FALSE,
  ssmax = FALSE,
  recompute = FALSE
)
}
\arguments{
\item{max_classes}{integer. Determines the task type and output behavior.
If 0: regression using quantile prediction. If > 0: classification with
hierarchical support if dataset classes exceed this value.}

\item{out_dim}{integer. Output dimension of the model.}

\item{d_model}{integer. Model dimension.}

\item{num_blocks}{integer. Number of blocks used in the ICL encoder.}

\item{nhead}{integer. Number of attention heads of the ICL encoder.}

\item{dim_feedforward}{integer. Dimension of the feedforward network.}

\item{dropout}{float. Dropout probability. Default 0.0.}

\item{activation}{str or function. Activation function ("relu", "gelu").}

\item{norm_first}{logical. If TRUE, uses pre-norm architecture.}

\item{bias_free_ln}{logical. If TRUE, removes bias from LayerNorm layers.}

\item{ssmax}{logical or str. Type of scalable softmax to use.}

\item{recompute}{logical. If TRUE, uses gradient checkpointing.}
}
\description{
Dataset-wise in-context learning.
}
