% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_icl2_predict.R
\name{predict.tab_icl_v2}
\alias{predict.tab_icl_v2}
\title{Predict using \code{TabICL2}}
\usage{
\method{predict}{tab_icl_v2}(object, new_data, ...)
}
\arguments{
\item{object, x}{A \code{tab_icl_v2} object.}

\item{new_data}{A data frame or matrix of new predictors.}

\item{...}{Not used, but required for extensibility.}
}
\value{
\code{\link[=predict]{predict()}} returns a tibble of predictions and \code{\link[=augment]{augment()}} appends the
columns in \code{new_data}. In either case, the number of rows in the tibble is
guaranteed to be the same as the number of rows in \code{new_data}.

For regression data, the prediction is in the column \code{.pred}. For
classification, the class predictions are in \code{.pred_class} and the
probability estimates are in columns with the pattern \verb{.pred_\{level\}} where
\code{level} is the levels of the outcome factor vector.
}
\description{
Predict using \code{TabICL2}
}
\examples{
# Minimal example for quick execution
car_split <- rsample::initial_split(mtcars[ 1:6,   ])

\dontrun{
# Fit
if (torch_is_installed() & interactive()) {
 mod <- tab_icl2(mpg ~ cyl + log(drat), car_split)

 # Predict
 predict(mod, testing(car_split))
 augment(mod, testing(car_split))
}
}

}
