% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rope.R
\name{rotary_embedding}
\alias{rotary_embedding}
\title{Rotary Positional Embedding for R {torch}}
\description{
Implementation of rotary positional embeddings (RoPE) for transformer models.
Ported from https://github.com/lucidrains/rotary-embedding-torch
}
\keyword{internal}
