% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rope.R
\name{RotaryEmbedding}
\alias{RotaryEmbedding}
\title{Rotary Positional Embeddings Module}
\usage{
RotaryEmbedding(
  dim,
  interleaved = TRUE,
  custom_freqs = NULL,
  freqs_for = "lang",
  theta = 10000,
  max_freq = 10,
  num_freqs = 1L,
  learned_freq = FALSE,
  use_xpos = FALSE,
  xpos_scale_base = 512,
  interpolate_factor = 1,
  theta_rescale_factor = 1,
  seq_before_head_dim = FALSE,
  cache_if_possible = TRUE
)
}
\arguments{
\item{dim}{Integer. The dimension of the embeddings.}

\item{interleaved}{Logical. If \code{TRUE}, uses interleaved rotation (default: \code{TRUE}).}

\item{custom_freqs}{Tensor or \code{NULL}. Custom frequency tensor (default: \code{NULL}).}

\item{freqs_for}{Character. One of \code{"lang"}, \code{"pixel"}, or \code{"constant"} (default: \code{"lang"}).}

\item{theta}{Numeric. Base scaling factor (default: \code{10000}).}

\item{max_freq}{Numeric. Maximum frequency for pixel-based embeddings (default: \code{10}).}

\item{num_freqs}{Integer. Number of frequencies for \code{"constant"} mode (default: \code{1L}).}

\item{learned_freq}{Logical. If \code{TRUE}, frequencies are learnable parameters (default: \code{FALSE}).}

\item{use_xpos}{Logical. If \code{TRUE}, uses extrapolatable rotary embeddings (XPOS) (default: \code{FALSE}).}

\item{xpos_scale_base}{Numeric. Base scaling factor used for XPOS (default: \code{512}).}

\item{interpolate_factor}{Numeric. Factor by which sequence length is interpolated (default: \code{1.0}).}

\item{theta_rescale_factor}{Numeric. Rescaling factor applied to theta (default: \code{1.0}).}

\item{seq_before_head_dim}{Logical. If \code{TRUE}, sequences are handled before head dim (default: \code{FALSE}).}

\item{cache_if_possible}{Logical. If \code{TRUE}, caches computed frequencies (default: \code{TRUE}).}
}
\value{
An \code{nn_module} object.
}
\description{
Rotary embeddings encode positional information in a way that allows
continuous rotation of embeddings, enhancing the model's ability to
capture long-range dependencies and positional relations.
}
