% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardhat_base.R
\name{runtime_error}
\alias{runtime_error}
\alias{type_error}
\alias{value_error}
\alias{stop_iteration_error}
\title{Custom error wrappers using glue-style formatting}
\usage{
runtime_error(msg, ...)

type_error(msg, ...)

value_error(msg, ...)

stop_iteration_error(msg, ...)
}
\arguments{
\item{msg}{Message template with {variable} placeholders}

\item{...}{Variables to interpolate in the message}
}
\description{
Custom error wrappers using glue-style formatting
}
\keyword{internal}
