% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardhat_base.R
\name{softmax}
\alias{softmax}
\title{Compute softmax values with temperature scaling}
\usage{
softmax(x, axis = -1, temperature = 0.9)
}
\arguments{
\item{x}{Input array of logits}

\item{axis}{Axis along which to compute softmax}

\item{temperature}{Temperature scaling parameter}
}
\value{
Softmax probabilities array
}
\description{
Computes softmax(x / temperature) using array operations.
}
