% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kv_cache.R
\docType{package}
\name{tabicl-cache}
\alias{TabICL2}
\alias{tabicl-cache}
\title{TabICL Key-Value Cache Structures}
\description{
Cache data structures for storing key-value projections from attention layers,
enabling efficient inference by reusing computed values across test samples.

The caching strategy focuses on:
\enumerate{
\item ColEmbedding: Cache K/V of the second attention layer of ISAB blocks
\item ICLearning: Cache K/V from training data at each layer of the ICL transformer
}
}
\author{
\strong{Maintainer}: Christophe Regouby \email{christophe.regouby@free.fr}

Authors:
\itemize{
  \item Christophe Regouby \email{christophe.regouby@free.fr}
}

}
\keyword{internal}
