% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_icl2_package.R
\docType{package}
\name{TabICL2-package}
\alias{TabICL2}
\alias{TabICL2-package}
\title{TabICL2: A TabICL2 tabular foundation model}
\description{
A TabICL2 tabular foundation model.
}
\author{
\strong{Maintainer}: Christophe Regouby \email{christophe.regouby@free.fr}

Authors:
\itemize{
  \item Christophe Regouby \email{christophe.regouby@free.fr}
}

}
\keyword{internal}
