% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kv_cache.R
\name{tabicl_cache_concat}
\alias{tabicl_cache_concat}
\title{Concatenate Multiple TabICLCache Objects}
\usage{
tabicl_cache_concat(caches, dim = 1)
}
\arguments{
\item{caches}{List of \code{TabICLCache} instances.}

\item{dim}{Integer. Dimension to concatenate along (batch dimension).
Default is \code{1} (R 1-based; corresponds to Python dim 0).}
}
\value{
A new \code{TabICLCache} with concatenated caches.
}
\description{
Concatenate multiple \code{TabICLCache} objects along the batch dimension.
}
