% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{torch_unique}
\alias{torch_unique}
\title{Get unique values from a tensor (pure torch implementation)}
\usage{
torch_unique(x)
}
\arguments{
\item{x}{torch tensor}
}
\value{
torch tensor with unique sorted values
}
\description{
Get unique values from a tensor (pure torch implementation)
}
\keyword{internal}
