% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{disk_tensor}
\alias{disk_tensor}
\title{A tensor backed by a memory-mapped file on disk}
\value{
A \code{DiskTensor} object.
}
\description{
Uses R's \code{bigmemory} or base file I/O for storage and provides a torch tensor
interface for seamless integration. This enables storing tensors larger than
available RAM by using disk space.
}
\note{
R6 doesn't support \code{[} overloading directly; use \verb{$get()} instead.
Write to the tensor (automatically persists to disk)
}
\section{Methods}{

\describe{
\item{\code{flush()}}{Flush changes to disk}
}
}

\examples{
\dontrun{
# Create a disk-backed tensor
dt <- disk_tensor$new(c(1000L, 128L), torch_float(), "/tmp/my_tensor.mmap")

# Write data (automatically persists to disk)
dt[1:100, ..] <- torch_randn(100L, 128L)

# Read data
x <- dt[1:50, ..]$tensor

# Flush changes to disk
dt$flush()
}

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DiskTensor-new}{\code{disk_tensor$new()}}
\item \href{#method-DiskTensor-tensor}{\code{disk_tensor$tensor()}}
\item \href{#method-DiskTensor-get}{\code{disk_tensor$get()}}
\item \href{#method-DiskTensor-set}{\code{disk_tensor$set()}}
\item \href{#method-DiskTensor-flush}{\code{disk_tensor$flush()}}
\item \href{#method-DiskTensor-nbytes}{\code{disk_tensor$nbytes()}}
\item \href{#method-DiskTensor-clone}{\code{disk_tensor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiskTensor-new"></a>}}
\if{latex}{\out{\hypertarget{method-DiskTensor-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{disk_tensor$new(shape, dtype, path, cleanup = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{shape}}{Integer vector. Shape of the tensor.}

\item{\code{dtype}}{torch.dtype. Data type.}

\item{\code{path}}{Character. File path.}

\item{\code{cleanup}}{Logical. Auto-cleanup on GC.
Get the torch tensor view (loads data if needed)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiskTensor-tensor"></a>}}
\if{latex}{\out{\hypertarget{method-DiskTensor-tensor}{}}}
\subsection{Method \code{tensor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{disk_tensor$tensor()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A torch tensor.
Index into the tensor (read)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiskTensor-get"></a>}}
\if{latex}{\out{\hypertarget{method-DiskTensor-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{disk_tensor$get(indices)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{indices}}{List of indices or slices.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A torch tensor.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiskTensor-set"></a>}}
\if{latex}{\out{\hypertarget{method-DiskTensor-set}{}}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{disk_tensor$set(indices, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{indices}}{List of indices.}

\item{\code{value}}{Tensor. Value to write.
Flush changes to disk}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiskTensor-flush"></a>}}
\if{latex}{\out{\hypertarget{method-DiskTensor-flush}{}}}
\subsection{Method \code{flush()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{disk_tensor$flush()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Invisible flush to disk
Get total size in bytes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiskTensor-nbytes"></a>}}
\if{latex}{\out{\hypertarget{method-DiskTensor-nbytes}{}}}
\subsection{Method \code{nbytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{disk_tensor$nbytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Integer. Size in bytes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiskTensor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiskTensor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{disk_tensor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
