% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kv-cache.R
\name{KVCache}
\alias{KVCache}
\title{KVCache -- Per-Layer Key-Value Cache}
\description{
An R6 class that maps layer/block indices
(\code{integer}) to \code{\link{KVCacheEntry}} objects.
Supports slicing, device transfer, concatenation and
pre-allocation.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KVCache-new}{\code{KVCache$new()}}
\item \href{#method-KVCache-is_populated}{\code{KVCache$is_populated()}}
\item \href{#method-KVCache-subset}{\code{KVCache$subset()}}
\item \href{#method-KVCache-assign}{\code{KVCache$assign()}}
\item \href{#method-KVCache-to}{\code{KVCache$to()}}
\item \href{#method-KVCache-preallocate}{\code{KVCache$preallocate()}}
\item \href{#method-KVCache-clone}{\code{KVCache$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KVCache-new"></a>}}
\if{latex}{\out{\hypertarget{method-KVCache-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KVCache$new(kv = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KVCache-is_populated"></a>}}
\if{latex}{\out{\hypertarget{method-KVCache-is_populated}{}}}
\subsection{Method \code{is_populated()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KVCache$is_populated()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KVCache-subset"></a>}}
\if{latex}{\out{\hypertarget{method-KVCache-subset}{}}}
\subsection{Method \code{subset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KVCache$subset(start, end)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KVCache-assign"></a>}}
\if{latex}{\out{\hypertarget{method-KVCache-assign}{}}}
\subsection{Method \code{assign()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KVCache$assign(start, end, other)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KVCache-to"></a>}}
\if{latex}{\out{\hypertarget{method-KVCache-to}{}}}
\subsection{Method \code{to()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KVCache$to(device, dtype = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KVCache-preallocate"></a>}}
\if{latex}{\out{\hypertarget{method-KVCache-preallocate}{}}}
\subsection{Method \code{preallocate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KVCache$preallocate(reference, batch_shape, device = "cpu", dtype = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KVCache-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KVCache-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KVCache$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
