% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kv-cache.R
\name{KVCacheEntry}
\alias{KVCacheEntry}
\title{KVCacheEntry -- Single Key-Value Cache Entry}
\description{
An R6 class holding optional cached key and value
projections for one attention layer.  Supports slicing, device
transfer, concatenation and in-place assignment.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KVCacheEntry-new}{\code{KVCacheEntry$new()}}
\item \href{#method-KVCacheEntry-is_valid}{\code{KVCacheEntry$is_valid()}}
\item \href{#method-KVCacheEntry-subset}{\code{KVCacheEntry$subset()}}
\item \href{#method-KVCacheEntry-assign}{\code{KVCacheEntry$assign()}}
\item \href{#method-KVCacheEntry-to}{\code{KVCacheEntry$to()}}
\item \href{#method-KVCacheEntry-clone}{\code{KVCacheEntry$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KVCacheEntry-new"></a>}}
\if{latex}{\out{\hypertarget{method-KVCacheEntry-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KVCacheEntry$new(key = NULL, value = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KVCacheEntry-is_valid"></a>}}
\if{latex}{\out{\hypertarget{method-KVCacheEntry-is_valid}{}}}
\subsection{Method \code{is_valid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KVCacheEntry$is_valid()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KVCacheEntry-subset"></a>}}
\if{latex}{\out{\hypertarget{method-KVCacheEntry-subset}{}}}
\subsection{Method \code{subset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KVCacheEntry$subset(start, end)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KVCacheEntry-assign"></a>}}
\if{latex}{\out{\hypertarget{method-KVCacheEntry-assign}{}}}
\subsection{Method \code{assign()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KVCacheEntry$assign(start, end, other)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KVCacheEntry-to"></a>}}
\if{latex}{\out{\hypertarget{method-KVCacheEntry-to}{}}}
\subsection{Method \code{to()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KVCacheEntry$to(device, dtype = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KVCacheEntry-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KVCacheEntry-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KVCacheEntry$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
