% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{pinned_buffer_pool}
\alias{pinned_buffer_pool}
\title{Pool of pinned CPU memory buffers for efficient GPU-to-CPU transfers}
\value{
A \code{PinnedBufferPool} object.
}
\description{
Pinned (page-locked) memory enables faster async GPU-to-CPU data transfers
by allowing the GPU to directly access the CPU memory without involving
the CPU. This class maintains a pool of such buffers to avoid the overhead
of repeated allocation and deallocation.
}
\examples{
\dontrun{
pool <- pinned_buffer_pool$new(max_buffers_per_shape = 8L)
buf <- pool$get(c(100L, 128L), torch_float())
# ... use buffer ...
pool$put(buf)  # Return to pool for reuse
}

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PinnedBufferPool-new}{\code{pinned_buffer_pool$new()}}
\item \href{#method-PinnedBufferPool-get}{\code{pinned_buffer_pool$get()}}
\item \href{#method-PinnedBufferPool-put}{\code{pinned_buffer_pool$put()}}
\item \href{#method-PinnedBufferPool-clear}{\code{pinned_buffer_pool$clear()}}
\item \href{#method-PinnedBufferPool-clone}{\code{pinned_buffer_pool$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PinnedBufferPool-new"></a>}}
\if{latex}{\out{\hypertarget{method-PinnedBufferPool-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pinned_buffer_pool$new(max_buffers_per_shape = 4L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{max_buffers_per_shape}}{Integer. Maximum buffers per (shape, dtype)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PinnedBufferPool-get"></a>}}
\if{latex}{\out{\hypertarget{method-PinnedBufferPool-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pinned_buffer_pool$get(shape, dtype)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{shape}}{Integer vector. Shape of the tensor.}

\item{\code{dtype}}{torch.dtype. Data type of the tensor.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A pinned CPU tensor.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PinnedBufferPool-put"></a>}}
\if{latex}{\out{\hypertarget{method-PinnedBufferPool-put}{}}}
\subsection{Method \code{put()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pinned_buffer_pool$put(buf)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{buf}}{Tensor. Buffer to return.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PinnedBufferPool-clear"></a>}}
\if{latex}{\out{\hypertarget{method-PinnedBufferPool-clear}{}}}
\subsection{Method \code{clear()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pinned_buffer_pool$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PinnedBufferPool-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PinnedBufferPool-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pinned_buffer_pool$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
