% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kv-cache.R
\name{tabicl_cache_concat}
\alias{tabicl_cache_concat}
\title{Concatenate Multiple TabICLCache Objects}
\usage{
tabicl_cache_concat(caches, dim = 1L)
}
\arguments{
\item{caches}{\code{list} of \code{TabICLCache} objects.}

\item{dim}{\code{integer(1)} Dimension to concatenate along
(batch dimension).  Default: \code{1L}.}
}
\value{
A new \code{TabICLCache}.
}
\description{
Concatenate the \code{col_cache}, \code{row_repr},
and \code{icl_cache} of several \code{\link{TabICLCache}}
objects along the batch dimension.
}
\seealso{
\code{\link{TabICLCache}}, \code{\link{KVCache}}
}
