% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kv-cache.R
\name{TabICLCache}
\alias{TabICLCache}
\title{TabICLCache -- Top-Level Cache for the TabICL Model}
\description{
Aggregates caches for the three major TabICL
components:
\itemize{
\item \strong{ColEmbedding} cache (ISAB blocks in column
embedding).
\item \strong{ICLearning} cache (Encoder layers in the ICL
transformer).
\item \strong{Row representations} cached as a single tensor.
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TabICLCache-new}{\code{TabICLCache$new()}}
\item \href{#method-TabICLCache-cache_type}{\code{TabICLCache$cache_type()}}
\item \href{#method-TabICLCache-cache_size_mb}{\code{TabICLCache$cache_size_mb()}}
\item \href{#method-TabICLCache-is_empty}{\code{TabICLCache$is_empty()}}
\item \href{#method-TabICLCache-slice_batch}{\code{TabICLCache$slice_batch()}}
\item \href{#method-TabICLCache-to}{\code{TabICLCache$to()}}
\item \href{#method-TabICLCache-clone}{\code{TabICLCache$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TabICLCache-new"></a>}}
\if{latex}{\out{\hypertarget{method-TabICLCache-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TabICLCache$new(
  col_cache = NULL,
  row_repr = NULL,
  icl_cache = NULL,
  train_shape = c(0L, 0L, 0L),
  num_classes = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TabICLCache-cache_type"></a>}}
\if{latex}{\out{\hypertarget{method-TabICLCache-cache_type}{}}}
\subsection{Method \code{cache_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TabICLCache$cache_type()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TabICLCache-cache_size_mb"></a>}}
\if{latex}{\out{\hypertarget{method-TabICLCache-cache_size_mb}{}}}
\subsection{Method \code{cache_size_mb()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TabICLCache$cache_size_mb()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TabICLCache-is_empty"></a>}}
\if{latex}{\out{\hypertarget{method-TabICLCache-is_empty}{}}}
\subsection{Method \code{is_empty()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TabICLCache$is_empty()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TabICLCache-slice_batch"></a>}}
\if{latex}{\out{\hypertarget{method-TabICLCache-slice_batch}{}}}
\subsection{Method \code{slice_batch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TabICLCache$slice_batch(start, end)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TabICLCache-to"></a>}}
\if{latex}{\out{\hypertarget{method-TabICLCache-to}{}}}
\subsection{Method \code{to()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TabICLCache$to(device, dtype = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TabICLCache-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TabICLCache-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TabICLCache$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
